Clean up configs
Removes obsolete TODO comments and allows the cluster name to be set
via the config map rather than hard-coded in the config template.

One FIXME to find a way to use the image specified in the config map
too, which currently appears to be unused. I'm not sure if there's
a way to retrieve the image value from the config map, but it seems
like there must be.