Add script to automate deployment
The script serves two purposes: First, it automates all of the
commands that were previously documented in the readme. Second, it
automates updating of the image in the pod definitions. It's not
possible to look up the image location from the config map directly,
but we can look at the config map YAML from the script and extract
the image location that way. This makes user_config.yaml the only
place users will need to modify values.